+++ Plugin/src/GiantSpecimensConfig.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using BepInEx;
 using BepInEx.Configuration;
+using GiantSpecimens.Scrap;
 
 namespace GiantSpecimens.Configs {
     public class GiantSpecimensConfig {
@@ -24,7 +25,32 @@ namespace GiantSpecimens.Configs {
         public ConfigEntry<bool> ConfigDriftWoodPlushieEnabled { get; private set; }
         public ConfigEntry<string> ConfigDriftWoodPlushieRarity { get; private set; }
         public ConfigEntry<int> ConfigMultiplierDriftwood { get; private set; }
+        public ConfigEntry<bool> ConfigZeusMode { get; private set; }
+        public ConfigEntry<bool> ConfigEatOldBirds { get; private set; }
+        public ConfigEntry<float> ConfigScreamRange { get; private set; }
+        public ConfigEntry<bool> ConfigDriftwoodHeartEnabled { get; private set; }
+        public ConfigEntry<bool> ConfigRedwoodHeartEnabled { get; private set; }
         public GiantSpecimensConfig(ConfigFile configFile) {
+            ConfigDriftwoodHeartEnabled = configFile.Bind("Scrap Options",
+                                                "DriftWood Giant | Heart",
+                                                true,
+                                                "Enables/Disables the Heart from spawning on death.");
+            ConfigRedwoodHeartEnabled = configFile.Bind("Scrap Options",
+                                                "RedWood Giant | Heart",
+                                                true,
+                                                "Enables/Disables the Heart from spawning on death.");
+            ConfigZeusMode = configFile.Bind("Enemy Options",
+                                                "RedWood Giant | Zeus Mode",
+                                                false,
+                                                "Enables/Disables the Zeus mode.");
+            ConfigEatOldBirds = configFile.Bind("Enemy Options",
+                                                "RedWood Giant | Eating Palette",
+                                                true,
+                                                "Enables/Disables the eating of old birds.");
+            ConfigScreamRange = configFile.Bind("Enemy Options",
+                                                "DriftWood Giant | Scream Range",
+                                                20f,
+                                                "Configure the range of the scream of the Driftwood Giant");
             ConfigMultiplierForestkeeper = configFile.Bind("Enemy Options",   // The section under which the option is shown
                                                 "ForestKeeper Multiplier",  // The key of the configuration option in the configuration file
                                                 2, // The default value
@@ -39,7 +65,7 @@ namespace GiantSpecimens.Configs {
                                                 "Enables/Disables the spawning of the RedWood Giant (sets rarity to 0 if false on all moons)"); // Description of the option to show in the config file
             ConfigRedWoodRarity = configFile.Bind("Enemy Options", 
                                                 "RedWood Giant | Spawn Weight.",
-                                                "Modded@100,Experimentation@50,Assurance@100,Vow@200,Offense@100,March@200,Rend@200,Dine@100,Titan@200",
+                                                "Modded@100,ExperimentationLevel@50,AssuranceLevel@100,VowLevel@200,OffenseLevel@100,MarchLevel@200,RendLevel@200,DineLevel@100,TitanLevel@200,Adamance@100,Embrion@150,Artifice@200",
                                                 "Spawn Weight of the RedWood Giant in all moons, Feel free to add to it any moon, just follow the format (also needs LLL installed for LE moons to work with this config).");
             ConfigSpeedRedWood = configFile.Bind("Enemy Options",   
                                                 "RedWood Giant Speed",  
@@ -63,7 +89,7 @@ namespace GiantSpecimens.Configs {
                                                 "Enables/Disables the spawning of the scrap (sets rarity to 0 if false on all moons)");
             ConfigWhistleRarity = configFile.Bind("Scrap Options",   
                                                 "Whistle Scrap | Rarity",  
-                                                "Modded@5,Experimentation@5,Assurance@5,Vow@5,Offense@5,March@5,Rend@5,Dine@5,Titan@5", 
+                                                "Modded@5,ExperimentationLevel@5,AssuranceLevel@5,VowLevel@5,OffenseLevel@5,MarchLevel@5,RendLevel@5,DineLevel@5,TitanLevel@5,Adamance@5,Embrion@5,Artifice@5", 
                                                 "Rarity of Whistle scrap appearing on every moon");
             ConfigRedwoodPlushieEnabled = configFile.Bind("Scrap Options",
                                                 "RedWood Giant Scrap | Enabled",
@@ -71,7 +97,7 @@ namespace GiantSpecimens.Configs {
                                                 "Enables/Disables the spawning of the scrap (sets rarity to 0 if false on all moons)");
             ConfigRedwoodPlushieRarity = configFile.Bind("Scrap Options",   
                                                 "RedWood Giant Scrap | Rarity",  
-                                                "Modded@5,Experimentation@5,Assurance@5,Vow@5,Offense@5,March@5,Rend@5,Dine@5,Titan@5", 
+                                                "Modded@5,ExperimentationLevel@5,AssuranceLevel@5,VowLevel@5,OffenseLevel@5,MarchLevel@5,RendLevel@5,DineLevel@5,TitanLevel@5,Adamance@5,Embrion@5,Artifice@5", 
                                                 "Rarity of redwood plushie appearing on every moon");
             ConfigWhistleEnabled = configFile.Bind("Shop Options",   
                                                 "Whistle Item | Enabled",  
@@ -91,11 +117,11 @@ namespace GiantSpecimens.Configs {
                                                             "Enables/Disables the spawning of the scrap (sets rarity to 0 if false on all moons)");
             ConfigDriftWoodRarity = configFile.Bind("Enemy Options",
                                                     "Driftwood | Rarity",
-                                                    "Modded@100,Experimentation@75,Assurance@50,Vow@150,Offense@50,March@175,Rend@125,Dine@125,Titan@150",
+                                                    "Modded@100,ExperimentationLevel@75,AssuranceLevel@50,VowLevel@150,OffenseLevel@50,MarchLevel@175,RendLevel@125,DineLevel@125,TitanLevel@150,Adamance@100,Embrion@150,Artifice@200",
                                                     "Rarity of driftwood appearing on every moon");
             ConfigDriftWoodPlushieRarity = configFile.Bind("Scrap Options",
                                                         "Driftwood Scrap | Rarity",
-                                                        "Modded@5,Experimentation5,Assurance5,Vow@5,Offense5,March@5,Rend@5,Dine@5,Titan@5",
+                                                        "Modded@5,ExperimentationLevel5,AssuranceLevel@5,VowLevel@5,OffenseLevel@5,MarchLevel@5,RendLevel@5,DineLevel@5,TitanLevel@5,Adamance@5,Embrion@5,Artifice@5",
                                                         "Rarity of driftwood plushie appearing on every moon.");
             ClearUnusedEntries(configFile);
             Plugin.Logger.LogInfo("Setting up config for Giant Specimen plugin...");